create adapter, fragments and other things for sliding of the rules + packages
