+++ app/src/main/java/com/example/spygame/rules/VpAdapter.kt
@@ -1,8 +1,8 @@
-package com.example.spygame.fragments_rules
+package com.example.spygame.rules
 
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
+import com.example.spygame.rules.fragments.SlideFragment
 
 class VpAdapter(fa: SlideFragment, private val list: List<Fragment>): FragmentStateAdapter(fa){
     override fun getItemCount(): Int {
