+++ app/src/main/res/layout/fragment_slide.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments_rules.SlideFragment">
+    tools:context=".rules.fragments.SlideFragment">
 
     <com.google.android.material.tabs.TabLayout
         android:id="@+id/tbLayout"
